+++ test/test.js
@@ -26,42 +26,6 @@ describe('typer.format(obj)', function () {
     assert.equal(str, 'image/svg+xml')
   })
 
-  it('should format type with parameter', function () {
-    var str = typer.format({
-      type: 'text',
-      subtype: 'html',
-      parameters: {charset: 'utf-8'}
-    })
-    assert.equal(str, 'text/html; charset=utf-8')
-  })
-
-  it('should format type with parameter that needs quotes', function () {
-    var str = typer.format({
-      type: 'text',
-      subtype: 'html',
-      parameters: {foo: 'bar or "baz"'}
-    })
-    assert.equal(str, 'text/html; foo="bar or \\"baz\\""')
-  })
-
-  it('should format type with parameter with empty value', function () {
-    var str = typer.format({
-      type: 'text',
-      subtype: 'html',
-      parameters: {foo: ''}
-    })
-    assert.equal(str, 'text/html; foo=""')
-  })
-
-  it('should format type with multiple parameters', function () {
-    var str = typer.format({
-      type: 'text',
-      subtype: 'html',
-      parameters: {charset: 'utf-8', foo: 'bar', bar: 'baz'}
-    })
-    assert.equal(str, 'text/html; bar=baz; charset=utf-8; foo=bar')
-  })
-
   it('should require argument', function () {
     assert.throws(typer.format.bind(null), /obj.*required/)
   })
@@ -91,16 +55,6 @@ describe('typer.format(obj)', function () {
     var obj = {type: 'image', subtype: 'svg', suffix: 'xml\\'}
     assert.throws(typer.format.bind(null, obj), /invalid suffix/)
   })
-
-  it('should reject invalid parameter name', function () {
-    var obj = {type: 'image', subtype: 'svg', parameters: {'foo/': 'bar'}}
-    assert.throws(typer.format.bind(null, obj), /invalid parameter name/)
-  })
-
-  it('should reject invalid parameter value', function () {
-    var obj = {type: 'image', subtype: 'svg', parameters: {'foo': 'bar\u0000'}}
-    assert.throws(typer.format.bind(null, obj), /invalid parameter value/)
-  })
 })
 
 describe('typer.parse(string)', function () {
@@ -117,26 +71,6 @@ describe('typer.parse(string)', function () {
     assert.equal(type.suffix, 'xml')
   })
 
-  it('should parse parameters', function () {
-    var type = typer.parse('text/html; charset=utf-8; foo=bar')
-    assert.equal(type.type, 'text')
-    assert.equal(type.subtype, 'html')
-    assert.deepEqual(type.parameters, {
-      charset: 'utf-8',
-      foo: 'bar'
-    })
-  })
-
-  it('should parse parameters with extra LWS', function () {
-    var type = typer.parse('text/html ; charset=utf-8 ; foo=bar')
-    assert.equal(type.type, 'text')
-    assert.equal(type.subtype, 'html')
-    assert.deepEqual(type.parameters, {
-      charset: 'utf-8',
-      foo: 'bar'
-    })
-  })
-
   it('should lower-case type', function () {
     var type = typer.parse('IMAGE/SVG+XML')
     assert.equal(type.type, 'image')
@@ -144,45 +78,10 @@ describe('typer.parse(string)', function () {
     assert.equal(type.suffix, 'xml')
   })
 
-  it('should lower-case parameter names', function () {
-    var type = typer.parse('text/html; Charset=UTF-8')
-    assert.deepEqual(type.parameters, {
-      charset: 'UTF-8'
-    })
-  })
-
-  it('should unquote parameter values', function () {
-    var type = typer.parse('text/html; charset="UTF-8"')
-    assert.deepEqual(type.parameters, {
-      charset: 'UTF-8'
-    })
-  })
-
-  it('should unquote parameter values with escapes', function () {
-    var type = typer.parse('text/html; charset = "UT\\F-\\\\\\"8\\""')
-    assert.deepEqual(type.parameters, {
-      charset: 'UTF-\\"8"'
-    })
-  })
-
-  it('should handle balanced quotes', function () {
-    var type = typer.parse('text/html; param="charset=\\"utf-8\\"; foo=bar"; bar=foo')
-    assert.deepEqual(type.parameters, {
-      param: 'charset="utf-8"; foo=bar',
-      bar: 'foo'
-    })
-  })
-
   invalidTypes.forEach(function (type) {
     it('should throw on invalid media type ' + type, function () {
       assert.throws(typer.parse.bind(null, type), /invalid media type/)
     })
-  })
-
-  it('should throw on invalid parameter format', function () {
-    assert.throws(typer.parse.bind(null, 'text/plain; foo="bar'), /invalid parameter format/)
-    assert.throws(typer.parse.bind(null, 'text/plain; profile=http://localhost; foo=bar'), /invalid parameter format/)
-    assert.throws(typer.parse.bind(null, 'text/plain; profile=http://localhost'), /invalid parameter format/)
   })
 
   it('should require argument', function () {