+++ HISTORY.md
@@ -1,9 +1,12 @@
 unreleased
 ==========
 
+This major release brings the module back to it's RFC 6838 roots. If you want
+a module to parse the `Content-Type` or similar HTTP headers, use the
+`content-type` module instead.
+
   * Remove parameter handling, which is outside RFC 6838 scope
   * Remove `parse(req)` and `parse(res)` signatures
-    - Use the `content-type` module for content type parsing
   * perf: use a class for object creation
 
 0.3.0 / 2014-09-07